// client/src/gen/broker/protocol/broker.ts
export interface HeaderEtcd {
  /**
   * cluster_id is the ID of the cluster.
   * @format uint64
   */
  clusterId?: string;

  /**
   * member_id is the ID of the member.
   * @format uint64
   */
  memberId?: string;

  /**
   * revision is the Etcd key-value store revision when the request was
   * applied.
   * @format int64
   */
  revision?: string;

  /**
   * raft_term is the raft term when the request was applied.
   * @format uint64
   */
  raftTerm?: string;
}
// ...
export interface ProcessSpecID {
  /**
   * "Zone" in which the process is running. Zones may be AWS, Azure, or
   * Google Cloud Platform zone identifiers, or rack locations within a colo,
   * or given some other custom meaning. Gazette will replicate across
   * multiple zones, and seeks to minimize traffic which must cross zones (for
   * example, by proxying reads to a broker in the current zone).
   */
  zone?: string;

  /**
   * Unique suffix of the process within |zone|. It is permissible for a
   * suffix value to repeat across zones, but never within zones. In practice,
   * it's recommended to use a FQDN, Kubernetes Pod name, or comparable unique
   * and self-describing value as the ID suffix.
   */
  suffix?: string;
}
// ...
export type ProtocolCompressionCodec =
  | "INVALID"
  | "NONE"
  | "GZIP"
  | "ZSTANDARD"
  | "SNAPPY"
  | "GZIP_OFFLOAD_DECOMPRESSION";

/**
* Fragment is a content-addressed description of a contiguous Journal span,
defined by the [begin, end) offset range covered by the Fragment and the
SHA1 sum of the corresponding Journal content.
*/
export interface ProtocolFragment {
  /** Journal of the Fragment. */
  journal?: string;

  /**
   * Begin (inclusive) and end (exclusive) offset of the Fragment within the
   * Journal.
   * @format int64
   */
  begin?: string;

  /** @format int64 */
  end?: string;

  /** SHA1 sum of the Fragment's content. */
  sum?: ProtocolSHA1Sum;

  /** Codec with which the Fragment's content is compressed. */
  compressionCodec?: ProtocolCompressionCodec;

  /**
   * Fragment store which backs the Fragment. Empty if the Fragment has yet to
   * be persisted and is still local to a Broker.
   */
  backingStore?: string;

  /**
   * Modification timestamp of the Fragment within the backing store,
   * represented as seconds since the epoch.
   * @format int64
   */
  modTime?: string;

  /**
   * Path postfix under which the fragment is persisted to the store.
   * The complete Fragment store path is built from any path components of the
   * backing store, followed by the journal name, followed by the path postfix.
   */
  pathPostfix?: string;
}
// ...
export interface ProtocolHeader {
  /**
   * ID of the process responsible for request processing. May be empty iff
   * Header is being used within a proxied request, and that request may be
   * dispatched to any member of the Route.
   */
  processId?: ProcessSpecID;

  /**
   * Route of processes specifically responsible for this RPC, or an empty Route
   * if any process is capable of serving the RPC.
   */
  route?: ProtocolRoute;

  /**
   * Etcd represents the effective Etcd MVCC state under which a Gazette broker
   * is operating in its processing of requests and responses. Its inclusion
   * allows brokers to reason about relative "happened before" Revision ordering
   * of apparent routing conflicts in proxied or replicated requests, as well
   * as enabling sanity checks over equality of Etcd ClusterId (and precluding,
   * for example, split-brain scenarios where different brokers are backed by
   * different Etcd clusters). Etcd is kept in sync with
   * etcdserverpb.ResponseHeader.
   */
  etcd?: HeaderEtcd;
}
// ...
export interface ProtocolReadResponse {
  /** Status of the Read RPC. */
  status?: ProtocolStatus;

  /**
   * Header of the response. Accompanies the first ReadResponse of the response
   * stream.
   */
  header?: ProtocolHeader;

  /**
   * The effective offset of the read. See ReadRequest offset.
   * @format int64
   */
  offset?: string;

  /**
   * The offset to next be written, by the next append transaction served by
   * broker. In other words, the last offset through which content is
   * available to be read from the Journal. This is a metadata field and will
   * not be returned with a content response.
   * @format int64
   */
  writeHead?: string;

  /**
   * Fragment to which the offset was mapped. This is a metadata field and will
   * not be returned with a content response.
   */
  fragment?: ProtocolFragment;

  /** If Fragment is remote, a URL from which it may be directly read. */
  fragmentUrl?: string;

  /**
   * Content chunks of the read.
   * @format byte
   */
  content?: string;
}
// ...
export interface ProtocolRoute {
  /** Members of the Route, ordered on ascending ProcessSpec.ID (zone, suffix). */
  members?: ProcessSpecID[];

  /**
   * Index of the ProcessSpec serving as primary within |members|,
   * or -1 of no member is currently primary.
   * @format int32
   */
  primary?: number;

  /**
   * Endpoints of each Route member. If not empty, |endpoints| has the same
   * length and order as |members|, and captures the endpoint of each one.
   */
  endpoints?: string[];
}

/**
 * SHA1Sum is a 160-bit SHA1 digest.
 */
export interface ProtocolSHA1Sum {
  /** @format uint64 */
  part1?: string;

  /** @format uint64 */
  part2?: string;

  /** @format int64 */
  part3?: number;
}
// ...
export type ProtocolStatus =
  | "OK"
  | "JOURNAL_NOT_FOUND"
  | "NO_JOURNAL_PRIMARY_BROKER"
  | "NOT_JOURNAL_PRIMARY_BROKER"
  | "NOT_JOURNAL_BROKER"
  | "INSUFFICIENT_JOURNAL_BROKERS"
  | "OFFSET_NOT_YET_AVAILABLE"
  | "WRONG_ROUTE"
  | "PROPOSAL_MISMATCH"
  | "ETCD_TRANSACTION_FAILED"
  | "NOT_ALLOWED"
  | "WRONG_APPEND_OFFSET"
  | "INDEX_HAS_GREATER_OFFSET"
  | "REGISTER_MISMATCH";